Restructure project around transceiver UI
