+++ bin/rx_timed_samples
