+++ bin/tx_samples_from_file
